==> 03_rf_all_xval.Rmd <==
---
title: "ALL chimerism dataset - xvalidation `ranger` with **mlr3**"
knit: (function(input_file, encoding) {
  out_dir <- 'docs';
  rmarkdown::render(input_file,
 encoding=encoding,
 output_file=file.path(dirname(input_file), out_dir, 'rf_all_xvalid_mlr3.html'))})
author:
  - David C. Shyr^[Stanford Medicine, dcshyr@stanford.edu]
  - Simon Brewer^[University of Utah, simon.brewer@geog.utah.edu]
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
    fig_caption: true
---

## Libraries

```{r message=FALSE, warning=FALSE}

set.seed(42)
library(tidyverse)
library(mlr3verse)

```


## Data

Read in data
```{r}

dat = read.csv("./data/all.all.df.csv")

```

Make outcome a binary variable (0/1 relapse)
```{r}

dat$rbin = factor(dat$rbin, levels = c("yes", "no"))

```

Filter out any tests that are post-relapse

```{r}

dat = dat[which(dat$bdate < dat$dor | is.na(dat$dor)), ]

```

Filter out relapse >720 days

```{r}

dat = dat[which(dat$rbin == "no" | dat$rtime < 720),]

```

Filter out any missing tests
```{r}

dat = dat[!is.na(dat$bmc_cdw) & !is.na(dat$bmc_cd3) & 
            !is.na(dat$bmc_cd15) & !is.na(dat$bmc_cd34) &
            !is.na(dat$pbc_cdw) & !is.na(dat$pbc_cd3) & 
            !is.na(dat$pbc_cd15) & !is.na(dat$pbc_cd34),]

```

Get $p(relapse)$ for baseline model

```{r}

prbin = sum(as.numeric(dat$rbin)-1) / nrow(dat)

```


```{r}

# dat2 <- dat %>%
#   select(rbin, txage, hla, tbi, abd, ci, mtx, mmf, agvhd, cgvhd,
#          bmc_cdw, bmc_cd3, bmc_cd15, bmc_cd34, 
#          pbc_cdw, pbc_cd3, pbc_cd15, pbc_cd34, ID)
dat2 <- dat %>%
  select(rbin, sex, txage, 
         rstatprtx, ghgp, tbi, 
         bmc_cdw, bmc_cd3, bmc_cd15, bmc_cd34, 
         pbc_cdw, pbc_cd3, pbc_cd15, pbc_cd34, ID)

dat2 <- dat2 %>% 
  mutate_if(is.character, as.factor)  %>% 
  mutate_if(is.integer, as.numeric) %>%
  # mutate(abd = tolower(abd)) %>%
  drop_na() %>%
  droplevels()

```

## Tuning

### Set up task

```{r}

task_chim <- TaskClassif$new(id = "all", backend = dat2, 
                             target = "rbin")

```

Define patients for use in cross validation

```{r}

# task_chim$col_roles$group <- "ID"
task_chim$set_col_roles("ID", remove_from = 'feature')

```

### Measures

```{r}

mlr_measures

```

```{r}

msr_acc <- msr("classif.acc")
msr_bacc <- msr("classif.bacc")
msr_auc <- msr("classif.auc")
msr_sens <- msr("classif.sensitivity")
msr_spec <- msr("classif.specificity")

```

### Resampler

```{r}

rsmp_xv <- rsmp("repeated_cv", folds = 5, repeats = 10)

```

### Learner

```{r}

lrn_rf <- lrn("classif.ranger",
              mtry.ratio = 0.788,
              sample.fraction = 0.8,
              num.trees = 500, 
              replace = FALSE,
              predict_type = "prob")

```   


## Threshold

```{r}

pop = po("learner_cv", lrn("classif.ranger", 
                           mtry.ratio = 0.778,
                           sample.fraction = 0.9,
                           num.trees = 500, 
                           replace = FALSE,
                           predict_type = "prob")) %>>%
  po("tunethreshold")

pop = po("learner_cv", lrn_rf) %>>%
  po("tunethreshold")

pop$train(task_chim)

pop$state

```

```{r}

gr = po(lrn_rf) %>>%
  po("threshold", param_vals = list(thresholds = 0.4))

```

## Cross-validation

```{r echo=FALSE}
lgr::get_logger("mlr3")$set_threshold("warn")
lgr::get_logger("bbotk")$set_threshold("warn")
```

```{r}

rr = resample(task_chim, lrn_rf, rsmp_xv, store_models = FALSE)
rr$aggregate(c(msr_acc, msr_bacc, msr_auc, msr_sens, msr_spec))

```

```{r}

rr = resample(task_chim, gr, rsmp_xv, store_models = FALSE)
rr$aggregate(c(msr_acc, msr_bacc, msr_auc, msr_sens, msr_spec))

```

## Cross-validation (baseline)

```{r}

lrn_base <- lrn("classif.featureless",
              predict_type = "prob")
rr = resample(task_chim, lrn_base, rsmp_xv, store_models = FALSE)
rr$aggregate(c(msr_acc, msr_bacc, msr_auc, msr_sens, msr_spec))

```
