==> 05_rf_all_prob_year.Rmd <==
---
title: "ALL dataset - rf plots (transplant year added)"
knit: (function(input_file, encoding) {
  out_dir <- 'docs';
  rmarkdown::render(input_file,
 encoding=encoding,
 output_file=file.path(dirname(input_file), out_dir, 'rf_all_year_predict.html'))})
author:
  - David C. Shyr^[Stanford Medicine, dcshyr@stanford.edu]
  - Simon Brewer^[University of Utah, simon.brewer@geog.utah.edu]
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
    fig_caption: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, dpi=300)
```

## Libraries

Set random seed for reproducibility
```{r}
set.seed(1234)
```

```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(lubridate)
library(ggpubr)
library(mlr3verse)
library(ranger)
library(vip)
library(pdp)
library(gridExtra)
```

## Data

Read in data
```{r}
all.df <- read.csv("./data/all.all.df.csv")
```

Convert dates
```{r}
all.df$dot <- ymd(all.df$dot)
all.df$dor <- ymd(all.df$dor)
all.df$bdate <- ymd(all.df$bdate)
all.df$pdate <- ymd(all.df$pdate)
## Get trasnplant year
all.df$tyear <- year(all.df$dot)
```

Convert all character strings to factors
```{r}
all.df <- all.df %>% mutate_if(is.character,as.factor)
```

Make outcome a binary variable (0/1 relapse)
```{r}
all.df$rbin <- factor(all.df$rbin, levels = c("yes", "no"))
```

Filter out any tests that are post-relapse
```{r}
all.df <- all.df[which(all.df$bdate < all.df$dor | is.na(all.df$dor)), ]
```

Filter out relapse >720 days
```{r}
all.df <- all.df[which(all.df$rbin == "no" | all.df$rtime < 720),]
```

Filter out any missing tests
```{r}
all.df <- all.df[!is.na(all.df$bmc_cdw) & !is.na(all.df$bmc_cd3) & 
                   !is.na(all.df$bmc_cd15) & !is.na(all.df$bmc_cd34) &
                   !is.na(all.df$pbc_cdw) & !is.na(all.df$pbc_cd3) & 
                   !is.na(all.df$pbc_cd15) & !is.na(all.df$pbc_cd34),]
all.df <<- all.df
```

```{r}

# dat2 <- dat %>%
#   select(rbin, txage, hla, tbi, abd, ci, mtx, mmf, agvhd, cgvhd,
#          bmc_cdw, bmc_cd3, bmc_cd15, bmc_cd34, 
#          pbc_cdw, pbc_cd3, pbc_cd15, pbc_cd34, ID)
all.df <- all.df %>%
  select(rbin, sex, txage, tyear, ## ADDED tyear
         rstatprtx, ghgp, agvhd, 
         bmc_cdw, bmc_cd3, bmc_cd15, bmc_cd34, 
         pbc_cdw, pbc_cd3, pbc_cd15, pbc_cd34, ID)

all.df <- all.df %>% 
  mutate_if(is.character, as.factor)  %>% 
  mutate_if(is.integer, as.numeric) %>%
  # mutate(abd = tolower(abd)) %>%
  drop_na() %>%
  droplevels()

```

## Random forest

### Set up task

```{r}

task_chim <- TaskClassif$new(id = "all", backend = all.df, 
                             target = "rbin")

```

Define patients for use in cross validation

```{r}

# task_chim$col_roles$group <- "ID"
task_chim$set_col_roles("ID", remove_from = 'feature')

```

### Learner

```{r}

lrn_rf <- lrn("classif.ranger",
              mtry.ratio = 0.89,
              sample.fraction = 0.9,
              num.trees = 1500, 
              replace = FALSE,
              importance = 'permutation',
              predict_type = "prob")

```   

### Training

```{r}

lrn_rf$train(task_chim)


```

How did it do?

```{r}
lrn_rf$model
```

## Variable importance

```{r}
vip(lrn_rf$model, num_features = 20, scale = TRUE)
```

## Partial plots

```{r warning=FALSE}

myvars <- c("txage", "tyear", "sex", 
            "rstatprtx", "ghgp", "agvhd",
            "bmc_cdw", "bmc_cd3", "bmc_cd15", "bmc_cd34",
            "pbc_cdw", "pbc_cd3", "pbc_cd15", "pbc_cd34")

for (i in myvars) {
  
  p1 <- pdp::partial(lrn_rf$model, 
                     train = task_chim$data(),
                     pred.var = i, 
                     prob = TRUE,
                     plot = TRUE, rug = TRUE, 
                     plot.engine = "ggplot2") + theme_light() + 
    ggtitle(paste("PDP:", i)) 
  print(p1)
}
```

### All peripheral blood lineaages

```{r}
pw <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = "pbc_cdw", prob = TRUE) 
p3 <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = "pbc_cd3", prob = TRUE) 
p15 <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = "pbc_cd15", prob = TRUE) 
p34 <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = "pbc_cd34", prob = TRUE) 

plot.df <- data.frame(test = c(pw$pbc_cdw, p3$pbc_cd3, p15$pbc_cd15, p34$pbc_cd34),
                      yhat = c(pw$yhat, p3$yhat, p15$yhat, p34$yhat),
                      type = c(rep("cdw", nrow(pw)),
                               rep("cd3", nrow(p3)),
                               rep("cd15", nrow(p15)),
                               rep("cd34", nrow(p34))
                      ))

ggline(plot.df, x = "test", y = "yhat", col = "type",
       numeric.x.axis = TRUE, 
       plot_type = "l", size = 1.5, main = "Chimerism values (PB, prob)")
```


### All bone marrow lineaages

```{r}
pw <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = "bmc_cdw", prob = TRUE) 
p3 <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = "bmc_cd3", prob = TRUE) 
p15 <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = "bmc_cd15", prob = TRUE) 
p34 <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = "bmc_cd34", prob = TRUE) 

plot.df <- data.frame(test = c(pw$bmc_cdw, p3$bmc_cd3, p15$bmc_cd15, p34$bmc_cd34),
                      yhat = c(pw$yhat, p3$yhat, p15$yhat, p34$yhat),
                      type = c(rep("cdw", nrow(pw)),
                               rep("cd3", nrow(p3)),
                               rep("cd15", nrow(p15)),
                               rep("cd34", nrow(p34))
                      ))

ggline(plot.df, x = "test", y = "yhat", col = "type", 
       numeric.x.axis = TRUE,
       plot_type = "l", size = 1.5, 
       main = "Chimerism values (BM, prob)")
```


## 2D plots

```{r}
pd <- pdp::partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = c("tyear", "txage"), chull = TRUE, prob = TRUE)
pdp1 <- plotPartial(pd, main = "TYear vs TAge")
print(pdp1)
```

```{r}
pd <- pdp::partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = c("tyear", "ghgp"), chull = TRUE, prob = TRUE)
p1 <- ggplot(pd, aes(x = tyear, y = yhat, col = ghgp)) + 
  geom_line(size = 1.5) + 
  ggtitle(paste("PDP:", i)) +
  theme_bw()
print(p1)
```

```{r warning=FALSE}
myvars <- c("sex", "rstatprtx", "ghgp")

for (i in myvars) {
  # Compute partial dependence data for lstat and rm
  pd <- pdp::partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = c("txage", i), chull = TRUE, prob = TRUE)
  
  # Default PDP
  # pdp1 <- plotPartial(pd, main = i)
  # print(pdp1)
  p1 <- ggplot(pd, aes_string(x = "txage", y = "yhat", col = i)) + 
    geom_line(size = 1.5) + 
    ggtitle(paste("PDP:", i)) +
    theme_bw()
  print(p1)
  
}
```

```{r warning=FALSE}
pd <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = c("bmc_cd3", "bmc_cd34"), chull = TRUE, prob = TRUE)
pdp1 <- plotPartial(pd, main = "Chimerism (BM, prob)")

pdp2 <- plotPartial(pd, levelplot = FALSE, zlab = "rbin", colorkey = TRUE,
                    screen = list(z = 160, x = -60))

grid.arrange(pdp1, pdp2, nrow = 1) 
print(pdp1)
```

```{r warning=FALSE}
pd <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = c("pbc_cd3", "pbc_cd34"), chull = TRUE, prob = TRUE)
pdp1 <- plotPartial(pd, main = "Chimerism (PB, prob)")

pdp2 <- plotPartial(pd, levelplot = FALSE, zlab = "rbin", colorkey = TRUE,
                    screen = list(z = 160, x = -60))

grid.arrange(pdp1, pdp2, nrow = 1) 
print(pdp1)
```

```{r eval=FALSE}
myvars <- c("bmc_cdw", "bmc_cd3", "bmc_cd15", "bmc_cd34",
            "pbc_cdw", "pbc_cd3", "pbc_cd15", "pbc_cd34")

for (i in myvars) {
  # Compute partial dependence data for lstat and rm
  pd <- partial(lrn_rf$model, 
              train = task_chim$data(), 
              pred.var = c("txage", i), chull = TRUE, prob = TRUE)
  
  # Default PDP
  pdp1 <- plotPartial(pd, main = i)
  
  pdp2 <- plotPartial(pd, levelplot = FALSE, zlab = "rbin", colorkey = TRUE,
                      screen = list(z = 160, x = -60))
  
  grid.arrange(pdp1, pdp2, nrow = 1) 
  
  print(pdp1)
  
}
```
